+++ NEWS.md
@@ -3,6 +3,7 @@
 #### New functions
 
 * `search_pv` added to send requests to the PatentsView API
-* Functions inside of `qry_funs` added to help users write queries
-* `get_fields` and `get_endpoints` added to quickly get possible field names and endpoints, respectively
-* `flatten_pv_data` added to flatten data results
+* `qry_funs` list add added with functions to help users write queries
+* `get_fields` and `get_endpoints` added to quickly get possible field names 
+  and endpoints, respectively
+* `unnest_pv_data` added to unnest the data frames in the returned data