+++ NEWS.md
@@ -3,7 +3,7 @@
 #### New functions
 
 * `search_pv` added to send requests to the PatentsView API
-* `qry_funs` list add added with functions to help users write queries
+* `qry_funs` list added with functions to help users write queries
 * `get_fields` and `get_endpoints` added to quickly get possible field names 
   and endpoints, respectively
 * `unnest_pv_data` added to unnest the data frames in the returned data